requirements.txt
tensorflow==2.3.0
streamlit==1.0.0
numpy==1.19.1
opencv_python==4.5.3.56
Pillow==7.2.0
